+++ src/resolution/frameworks/index.ts
@@ -24,6 +24,7 @@ import { redwoodResolver } from './redwood';
 import { angularResolver } from './angular';
 import { analogResolver } from './analog';
 import { solidRouterResolver } from './solid-router';
+import { solidStartResolver } from './solid-start';
 import { djangoResolver, flaskResolver, fastapiResolver } from './python';
 import { railsResolver } from './ruby';
 import { springResolver } from './java';
@@ -71,6 +72,7 @@ const FRAMEWORK_RESOLVERS: FrameworkResolver[] = [
   angularResolver,
   analogResolver,
   solidRouterResolver,
+  solidStartResolver,
   // Python
   djangoResolver,
   flaskResolver,